requirements.txt
bs4
openai
requests
python-docx
python-dotenv